+++ baker/formatWordpressPost.tsx
@@ -24,13 +24,78 @@ import { countryProfileSpecs } from "site/countryProfileProjects"
 import { formatGlossaryTerms } from "site/formatGlossary"
 import { getMutableGlossary, glossary } from "site/glossary"
 import { DataToken } from "site/DataToken"
-import {
-    extractLatex,
-    formatLatex,
-    INTERACTIVE_ICON_SVG,
-    DEEP_LINK_CLASS,
-    getHtmlContentWithStyles,
-} from "site/formatting"
+import { DEEP_LINK_CLASS, getHtmlContentWithStyles } from "site/formatting"
+import { mathjax } from "mathjax-full/js/mathjax"
+import { TeX } from "mathjax-full/js/input/tex"
+import { SVG } from "mathjax-full/js/output/svg"
+import { liteAdaptor } from "mathjax-full/js/adaptors/liteAdaptor"
+import { RegisterHTMLHandler } from "mathjax-full/js/handlers/html"
+import { AllPackages } from "mathjax-full/js/input/tex/AllPackages"
+
+const initMathJax = () => {
+    const adaptor = liteAdaptor()
+    RegisterHTMLHandler(adaptor)
+
+    const tex = new TeX({ packages: AllPackages })
+    const svg = new SVG({ fontCache: "none" })
+    const doc = mathjax.document("", {
+        InputJax: tex,
+        OutputJax: svg,
+    })
+
+    return function format(latex: string): string {
+        const node = doc.convert(latex, {
+            display: true,
+        })
+        return adaptor.outerHTML(node) // as HTML
+    }
+}
+
+const formatMathJax = initMathJax()
+
+// A modifed FontAwesome icon
+const INTERACTIVE_ICON_SVG = `<svg aria-hidden="true" focusable="false" data-prefix="fas" data-icon="hand-pointer" class="svg-inline--fa fa-hand-pointer fa-w-14" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 617">
+    <path fill="currentColor" d="M448,344.59v96a40.36,40.36,0,0,1-1.06,9.16l-32,136A40,40,0,0,1,376,616.59H168a40,40,0,0,1-32.35-16.47l-128-176a40,40,0,0,1,64.7-47.06L104,420.58v-276a40,40,0,0,1,80,0v200h8v-40a40,40,0,1,1,80,0v40h8v-24a40,40,0,1,1,80,0v24h8a40,40,0,1,1,80,0Zm-256,80h-8v96h8Zm88,0h-8v96h8Zm88,0h-8v96h8Z" transform="translate(0 -0.41)"/>
+    <path fill="currentColor" opacity="0.6" d="M239.76,234.78A27.5,27.5,0,0,1,217,192a87.76,87.76,0,1,0-145.9,0A27.5,27.5,0,1,1,25.37,222.6,142.17,142.17,0,0,1,1.24,143.17C1.24,64.45,65.28.41,144,.41s142.76,64,142.76,142.76a142.17,142.17,0,0,1-24.13,79.43A27.47,27.47,0,0,1,239.76,234.78Z" transform="translate(0 -0.41)"/>
+</svg>`
+
+const extractLatex = (html: string): [string, string[]] => {
+    const latexBlocks: string[] = []
+    html = html.replace(/\[latex\]([\s\S]*?)\[\/latex\]/gm, (_, latex) => {
+        latexBlocks.push(
+            latex.replace("\\[", "").replace("\\]", "").replace(/\$\$/g, "")
+        )
+        return "[latex]"
+    })
+    return [html, latexBlocks]
+}
+
+const formatLatex = async (
+    html: string,
+    latexBlocks?: string[]
+): Promise<string> => {
+    if (!latexBlocks) [html, latexBlocks] = extractLatex(html)
+
+    // return early so we don't do unnecessary work for sites without latex
+    if (!latexBlocks.length) return html
+
+    const compiled: string[] = []
+    for (const latex of latexBlocks) {
+        try {
+            compiled.push(formatMathJax(latex))
+        } catch (err) {
+            compiled.push(
+                `${latex} (Could not format equation due to MathJax error)`
+            )
+        }
+    }
+
+    let i = -1
+    return html.replace(/\[latex\]/g, () => {
+        i += 1
+        return compiled[i]
+    })
+}
 
 export const formatWordpressPost = async (
     post: FullPost,
