+++ baker/formatting.tsx
@@ -1,0 +1,74 @@
+import * as cheerio from "cheerio"
+import { FormattedPost, FormattingOptions } from "clientUtils/owidTypes"
+import { Country } from "clientUtils/countries"
+import { countryProfileDefaultCountryPlaceholder } from "site/countryProfileProjects"
+
+export const DEEP_LINK_CLASS = "deep-link"
+
+export const getHtmlContentWithStyles = (cheerEl: CheerioStatic) => {
+    // Inline styling
+    // Get the first root level <style> tag within the content as it gets
+    // stripped out by $("body").html() below. Voluntarily limits to 1 as there
+    // should not be a need for more.
+    const style =
+        cheerEl("style").length === 1
+            ? `<style>${cheerEl("style").html()}</style>`
+            : ""
+
+    // This is effectively a hack within a hack, as style tags are technically
+    // not allowed in the body (of the main article)
+    return `${style}${cheerEl("body").html()}`
+}
+
+export const extractFormattingOptions = (html: string): FormattingOptions => {
+    const formattingOptionsMatch = html.match(
+        /<!--\s*formatting-options\s+(.*)\s*-->/
+    )
+    return formattingOptionsMatch
+        ? parseFormattingOptions(formattingOptionsMatch[1])
+        : {}
+}
+
+// Converts "toc:false raw somekey:somevalue" to { toc: false, raw: true, somekey: "somevalue" }
+// If only the key is specified, the value is assumed to be true (e.g. "raw" above)
+const parseFormattingOptions = (text: string): FormattingOptions => {
+    const options: { [key: string]: string | boolean } = {}
+    text.split(/\s+/)
+        // filter out empty strings
+        .filter((s) => s && s.length > 0)
+        // populate options object
+        .forEach((option: string) => {
+            const [name, value] = option.split(":") as [
+                string,
+                string | undefined
+            ]
+            let parsedValue
+            if (value === undefined || value === "true") parsedValue = true
+            else if (value === "false") parsedValue = false
+            else parsedValue = value
+            options[name] = parsedValue
+        })
+    return options
+}
+
+export const formatCountryProfile = (
+    post: FormattedPost,
+    country: Country
+): FormattedPost => {
+    // Localize country selector
+    const htmlWithLocalizedCountrySelector = post.html.replace(
+        countryProfileDefaultCountryPlaceholder,
+        country.code
+    )
+
+    const cheerioEl = cheerio.load(htmlWithLocalizedCountrySelector)
+
+    // Inject country names on h3 headings which have been already identified as subsections
+    // (filtering them out based on whether they have a deep link anchor attached to them)
+    cheerioEl(`h3 a.${DEEP_LINK_CLASS}`).each((_, deepLinkAnchor) => {
+        const $deepLinkAnchor = cheerioEl(deepLinkAnchor)
+        $deepLinkAnchor.after(`${country.name}: `)
+    })
+
+    return { ...post, html: getHtmlContentWithStyles(cheerioEl) }
+}
